tweak release script

+++ release.sh
@@ -11,21 +11,26 @@ if [[ $docker_compose_migration_version != $migration_version ]]; then
   echo "The migration version in the docker-compose file is not the same as the one in package.json ($docker_compose_migration_version != $migration_version)"
   exit
 fi
+
 echo "(Re)bulding docker containers..."
 docker-compose down -v
 docker-compose build
 docker-compose up -d
 
 # wait a bit for graph-node to come (it will redirect with a 302)
-echo "wating for subgraph to start"
+echo "waiting for subgraph to start"
 while [[ "$(curl -s -o /dev/null -w ''%{http_code}'' 127.0.0.1:8000)" != "200" ]]; do sleep 5; done
 
-echo "moving .env to make sure we have default settings"
-mv .env .env_backup
+if [ -f .env ]; then
+  echo "moving .env to make sure we have default settings"
+  mv .env .env_backup
+fi
 echo "deploying subgraph"
 npm run deploy
-echo "restoring .env file"
-mv .env .env_backup
+if [ -f .env_backup ]; then
+  echo "restoring .env file"
+  mv .env_backup .env
+fi
 
 
 # commit the postgres image
